+++ src/workers/helper.py
@@ -1,11 +1,27 @@
-from datetime import datetime
+from datetime import datetime, timezone, timedelta
+from dataclasses import dataclass
+import logging
+import asyncio
 import discord
+from discord.utils import snowflake_time
 from discord.ext.commands import Bot
 from settings import get_settings
 from util import create_interaction_embed_context
 from typing import Optional
 
+
+class UnableToDM(Exception):
+    pass
+
+
+@dataclass
+class AutomodResults:
+    num_removed: int = 0
+    num_errors: int = 0
+
+
 settings = get_settings()
+logger = logging.getLogger(__name__)
 
 
 def create_autounexile_embed(
@@ -27,3 +43,58 @@ def create_automod_embed(self, channel_id, num_removed, num_error, timestamp: da
         timestamp=timestamp,
         description=f"Successfully removed {num_removed} inactive thread(s) from <#{channel_id}>.\n{num_error} inactive thread(s) failed to be removed.",
     )
+
+
+async def automod_thread(
+    self, channel_id, thread: discord.Thread, duration: int, num_result: AutomodResults
+):
+    if thread.flags.pinned:
+        # skip the for loop if the thread is pinned
+        return
+
+    # check if starter message was deleted
+    starter_message = None
+    try:
+        starter_message = await thread.fetch_message(thread.id)
+        await asyncio.sleep(0.5)
+    except discord.NotFound:
+        pass
+    except Exception as e:
+        logger.error(e)
+
+    now = datetime.now(timezone.utc)
+    last_post = thread.last_message_id
+    time_since = now - snowflake_time(last_post)
+    if starter_message is not None and time_since < timedelta(days=duration):
+        return
+
+    # delete thread and try to send DM to user
+    try:
+        await thread.delete()
+        logger.info(f"Thread {thread.id} has been deleted successfully")
+        num_result.num_removed += 1
+        # TODO: uncomment DM portion and sleep when backlog is dealt with
+        # if thread.owner is not None:
+        #     try:
+        #         await send_dm(
+        #             thread.owner,
+        #             f'Your thread "{thread.name}" in <#{channel_id}> has been automatically deleted as {duration} days have passed without any activity or the starter message has been deleted.',
+        #         )
+        #     except discord.Forbidden:
+        #         raise UnableToDM("User has DMs disabled")
+        # else:
+        #     raise UnableToDM("User is not in the server")
+    except UnableToDM as e:
+        logger.info(f"Unable to DM user {thread.owner_id}: {e}")
+        channel = self.get_channel(settings.logging_channel_id)
+        async with create_interaction_embed_context(
+            channel,
+            user=thread.owner,
+            timestamp=datetime.now(timezone.utc),
+            description=f"<@{thread.owner_id}>'s thread was deleted in <#{channel_id}> but a DM could not be sent: {e}",
+        ):
+            pass
+    except Exception as e:
+        logger.error(e)
+        num_result.num_errors += 1
+    # await asyncio.sleep(300)
